+++ helper-persistence/src/main/java/org/example/persistence/comment/entity/ReplyJpaEntity.java
@@ -1,4 +1,4 @@
-package org.example.persistence.post.entity;
+package org.example.persistence.comment.entity;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
@@ -16,25 +16,28 @@ import org.example.persistence.student.entity.StudentJpaEntity;
 import org.hibernate.annotations.ColumnDefault;
 
 @Getter
-@Entity(name = "comment")
+@Entity(name = "reply")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class CommentJpaEntity {
+public class ReplyJpaEntity {
 
     @Id
     @Column(nullable = false, columnDefinition = "BINARY(16)")
-    private UUID commentId;
+    private UUID replyId;
 
     @ManyToOne(cascade = CascadeType.ALL, optional = false, targetEntity = StudentJpaEntity.class)
     @JoinColumn(name = "studentId", referencedColumnName = "studentId")
     private StudentJpaEntity student;
 
-    @ManyToOne(cascade = CascadeType.ALL, optional = false, targetEntity = PostJpaEntity.class)
-    @JoinColumn(name = "postId", referencedColumnName = "postId")
-    private PostJpaEntity post;
+    @ManyToOne(cascade = CascadeType.ALL, optional = false, targetEntity = CommentJpaEntity.class)
+    @JoinColumn(name = "commentId", referencedColumnName = "commentId")
+    private CommentJpaEntity comment;
 
     @Column(nullable = false, columnDefinition = "TEXT")
     private String content;
+
+    @Column(nullable = true, columnDefinition = "VARCHAR(10)")
+    private String mention;
 
     @ColumnDefault("NOW()")
     @Column(nullable = false, columnDefinition = "DATETIME")
