+++ helper-persistence/src/main/java/org/example/persistence/comment/repository/CommentJpaRepository.java
@@ -1,7 +1,7 @@
-package org.example.persistence.post.repository;
+package org.example.persistence.comment.repository;
 
 import java.util.UUID;
-import org.example.persistence.post.entity.CommentJpaEntity;
+import org.example.persistence.comment.entity.CommentJpaEntity;
 import org.springframework.data.repository.CrudRepository;
 
 public interface CommentJpaRepository extends CrudRepository<CommentJpaEntity, UUID> {
