+++ helper-persistence/src/main/java/org/example/persistence/comment/repository/ReplyJpaRepository.java
@@ -1,7 +1,7 @@
-package org.example.persistence.post.repository;
+package org.example.persistence.comment.repository;
 
 import java.util.UUID;
-import org.example.persistence.post.entity.ReplyJpaEntity;
+import org.example.persistence.comment.entity.ReplyJpaEntity;
 import org.springframework.data.repository.CrudRepository;
 
 public interface ReplyJpaRepository extends CrudRepository<ReplyJpaEntity, UUID> {
